feat: auto-generate CLI help docs from Clap definitions
Split main.rs into lib.rs + main.rs so the Cli type is reusable.
Added clap-markdown to generate docs/HELP.md from Clap definitions,
with an example binary for regeneration and an integration test + CI
job to catch drift.

- src/lib.rs: all types and modules (public)
- src/main.rs: thin entrypoint
- examples/generate_help.rs: prints help markdown to stdout
- tests/check_help.rs: verifies docs/HELP.md matches Clap definitions
- CI: dedicated "CLI Help Docs" job using diff pattern
- README: links to docs/HELP.md instead of inline usage

Regenerate with: cargo run --example generate_help > docs/HELP.md
