+++ examples/generate_help.rs
@@ -1,0 +1,4 @@
+fn main() {
+    let help = clap_markdown::help_markdown::<detail_cli::Cli>();
+    print!("{help}");
+}
